nueva funcionalidades en algunos modelos
Añadimos nuevas funcionalidades a los modelos Diet y Menu.
Eliminamos los modelos antiguos.
Corrección de refrescar las listas al eliminar.
