+++ AlimentApp/src/main/java/dad/alimentapp/controllers/ManageDietController.java
@@ -6,12 +6,12 @@ import java.util.ResourceBundle;
 
 import dad.alimentapp.main.App;
 import dad.alimentapp.models.ControlDietMenu;
+import dad.alimentapp.models.Diet;
+import dad.alimentapp.models.Menu;
 import dad.alimentapp.models.MomentDay;
 import dad.alimentapp.models.Product;
+import dad.alimentapp.models.ProductMomentDay;
 import dad.alimentapp.models.Profile;
-import dad.alimentapp.models.app.Diet;
-import dad.alimentapp.models.app.Menu;
-import dad.alimentapp.models.app.ProductMomentDay;
 import dad.alimentapp.service.DietService;
 import dad.alimentapp.service.MenuService;
 import dad.alimentapp.service.ProductService;
@@ -37,12 +37,14 @@ import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 
 public class ManageDietController implements Initializable {
-	
+
 	/**
-	 * Realizamos el controlador que permitirá implementar la interfaz de Gestionar Dietas y darle toda la funcionalidad que necesita. 
+	 * Realizamos el controlador que permitirá implementar la interfaz de Gestionar
+	 * Dietas y darle toda la funcionalidad que necesita.
+	 * 
 	 * @author Andy
 	 */
-	
+
 	@FXML
 	private HBox view;
 
@@ -123,7 +125,7 @@ public class ManageDietController implements Initializable {
 	// MODEL
 	private static ListProperty<Menu> menus = new SimpleListProperty<>(FXCollections.observableArrayList());
 	private ObjectProperty<Menu> menuSelected = new SimpleObjectProperty<>();
-	
+
 	private static ListProperty<Diet> diets = new SimpleListProperty<>(FXCollections.observableArrayList());
 	private ObjectProperty<Diet> dietSelected = new SimpleObjectProperty<>();
 
@@ -146,23 +148,23 @@ public class ManageDietController implements Initializable {
 	public void initialize(URL location, ResourceBundle resources) {
 		// BINDINGS
 		loadDietsAndMenus();
-		
+
 		menuList.itemsProperty().bindBidirectional(menus);
 		dietList.itemsProperty().bindBidirectional(diets);
-		
+
 		menuSelected.bind(menuList.getSelectionModel().selectedItemProperty());
 		dietSelected.bind(dietList.getSelectionModel().selectedItemProperty());
-		
+
 		viewMenuButton.disableProperty().bind(menuList.getSelectionModel().selectedItemProperty().isNull());
 		modifyMenuButton.disableProperty().bind(menuList.getSelectionModel().selectedItemProperty().isNull());
 		removeMenuButton.disableProperty().bind(menuList.getSelectionModel().selectedItemProperty().isNull());
-		
+
 		modifyDietButton.disableProperty().bind(dietList.getSelectionModel().selectedItemProperty().isNull());
 		removeDietButton.disableProperty().bind(dietList.getSelectionModel().selectedItemProperty().isNull());
 	}
-	
-	public static void loadDietsAndMenus() {		
-		Profile profile = MainController.getProfileSelected();		
+
+	public static void loadDietsAndMenus() {
+		Profile profile = MainController.getProfileSelected();
 		diets.setAll(DietService.getAllDiets(profile));
 		menus.setAll(MenuService.getAllMenus(profile));
 	}
@@ -211,37 +213,35 @@ public class ManageDietController implements Initializable {
 
 	@FXML
 	void onRemoveDietButtonAction(ActionEvent event) {
-		
 		DietService.deleteDiet(dietSelected.get());
-		Profile profile = MainController.getProfileSelected();	
+		Profile profile = MainController.getProfileSelected();
 		diets.setAll(DietService.getAllDiets(profile));
-
+		ManageDietController.loadDietsAndMenus();
 	}
 
 	@FXML
 	void onRemoveMenuButtonAction(ActionEvent event) {
-		
 		MenuService.deleteMenu(menuSelected.get());
-		Profile profile = MainController.getProfileSelected();		
+		Profile profile = MainController.getProfileSelected();
 		menus.setAll(MenuService.getAllMenus(profile));
 		resetProductsMenu();
-
+		ManageDietController.loadDietsAndMenus();
 	}
 
 	private void resetProductsMenu() {
-		
+
 		ObservableList<Product> clearList = FXCollections.observableArrayList();
 		breakfastListView.setItems(clearList);
 		midMorningListView.setItems(clearList);
 		lunchListView.setItems(clearList);
 		snackListView.setItems(clearList);
 		dinnerListView.setItems(clearList);
-		
+
 	}
 
 	@FXML
 	void onViewMenuButtonAction(ActionEvent event) {
-		
+
 		loadProductsMenu();
 		breakfastListView.setItems(breakfastProductList.getProducts());
 		midMorningListView.setItems(midMorningProductList.getProducts());
@@ -250,12 +250,16 @@ public class ManageDietController implements Initializable {
 		dinnerListView.setItems(dinnerProductList.getProducts());
 
 	}
-	
+
 	private void loadProductsMenu() {
-		breakfastProductList = ProductService.getAllProductsToMenuOfMomentDay(menuSelected.get().getId(), MomentDay.DESAYUNO);
-		midMorningProductList = ProductService.getAllProductsToMenuOfMomentDay(menuSelected.get().getId(), MomentDay.MEDIA_MAÑANA);
-		lunchProductList = ProductService.getAllProductsToMenuOfMomentDay(menuSelected.get().getId(), MomentDay.ALMUERZO);
-		snackProductList = ProductService.getAllProductsToMenuOfMomentDay(menuSelected.get().getId(), MomentDay.MERIENDA);
+		breakfastProductList = ProductService.getAllProductsToMenuOfMomentDay(menuSelected.get().getId(),
+				MomentDay.DESAYUNO);
+		midMorningProductList = ProductService.getAllProductsToMenuOfMomentDay(menuSelected.get().getId(),
+				MomentDay.MEDIA_MAÑANA);
+		lunchProductList = ProductService.getAllProductsToMenuOfMomentDay(menuSelected.get().getId(),
+				MomentDay.ALMUERZO);
+		snackProductList = ProductService.getAllProductsToMenuOfMomentDay(menuSelected.get().getId(),
+				MomentDay.MERIENDA);
 		dinnerProductList = ProductService.getAllProductsToMenuOfMomentDay(menuSelected.get().getId(), MomentDay.CENA);
 	}
 
@@ -288,11 +292,11 @@ public class ManageDietController implements Initializable {
 	public static Stage getChoiceStage() {
 		return choiceStage;
 	}
-	
+
 	public static Stage getModificateStage() {
 		return modifIcateStage;
 	}
-	
+
 	public static void setModificateStage(Stage stage) {
 		modifIcateStage = stage;
 	}
