+++ AlimentApp/src/main/java/dad/alimentapp/models/ProductMomentDay.java
@@ -1,9 +1,7 @@
-package dad.alimentapp.models.app;
+package dad.alimentapp.models;
 
 import java.util.List;
 
-import dad.alimentapp.models.MomentDay;
-import dad.alimentapp.models.Product;
 import javafx.beans.property.ListProperty;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleListProperty;
