+++ AlimentApp/src/main/java/dad/alimentapp/service/MenuService.java
@@ -8,15 +8,14 @@ import java.util.ArrayList;
 import java.util.List;
 
 import dad.alimentapp.main.App;
+import dad.alimentapp.models.Menu;
 import dad.alimentapp.models.MomentDay;
 import dad.alimentapp.models.Product;
+import dad.alimentapp.models.ProductMomentDay;
 import dad.alimentapp.models.Profile;
-import dad.alimentapp.models.app.Menu;
-import dad.alimentapp.models.app.ProductMomentDay;
 import dad.alimentapp.utils.Messages;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
-
 
 /**
  * En esta clase tenemos almacenadas todas las consultas a la base de datos, en
@@ -191,7 +190,7 @@ public class MenuService {
 			Messages.error("Error al insertar el nuevo menú", e.getMessage());
 		}
 	}
-	
+
 	/**
 	 * Esta función deleteMenu, nos permite eliminar un menú y sus claves foráneas
 	 * 
@@ -219,5 +218,5 @@ public class MenuService {
 			Messages.error("Error al eliminar el menú", e.getMessage());
 		}
 	}
-	
+
 }
